feat(web): rebrand Knowledge as Guideline and add Agents nav
Group New Request, Automations, and a new AI Engine page under an Agents sidebar section, and move model provider settings off Integrations.
