+++ apps/hyperlocalise-web/src/api/auth/LOCALIZATION_ROLES.md
@@ -10,7 +10,7 @@ reconcile, webhooks, and member mutations.
 | WorkOS / local slug    | Purpose                                                                                                                        |
 | ---------------------- | ------------------------------------------------------------------------------------------------------------------------------ |
 | `admin`                | Full workspace control including billing.                                                                                      |
-| `localization_manager` | Operate projects, integrations, credentials, teams, and knowledge resources; approve reviews and write-back. No billing write. |
+| `localization_manager` | Operate projects, integrations, credentials, teams, and guideline resources; approve reviews and write-back. No billing write. |
 | `developer`            | Manage projects and technical jobs (sync, repos); read integrations. No review approval, credentials, members, or billing.     |
 | `reviewer`             | Contribute to jobs, run AI actions, push draft translations; approve reviews and write-back. No org administration.            |
 | `translator`           | Contribute to assigned jobs, run AI actions, push draft translations. No approvals or org administration.                      |
