+++ apps/hyperlocalise-web/src/app/[lang]/(authenticated)/org/[organizationSlug]/ai-engine/page.tsx
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2026 Hyperlocalise Pty Ltd
+ *
+ * Use of this software is governed by the Business Source License 1.1
+ * included in this application's LICENSE file.
+ *
+ * Change Date: Four years after publication of the applicable version.
+ *
+ * On the Change Date, in accordance with the Business Source License, use
+ * of this software will be governed by the GNU General Public License
+ * Version 2.0 or later.
+ */
+import { hasCapability } from "@/api/auth/policy";
+import { requireAppAuthContext } from "@/lib/workos/app-auth";
+
+import { AiEnginePageContent } from "./_components/ai-engine-page-content";
+
+export default async function AiEnginePage({
+  params,
+}: {
+  params: Promise<{ organizationSlug: string }>;
+}) {
+  const { organizationSlug } = await params;
+  const auth = await requireAppAuthContext({ organizationSlug });
+
+  return (
+    <AiEnginePageContent
+      organizationSlug={organizationSlug}
+      canManageAiEngine={hasCapability(auth.membership.role, "integrations:read")}
+    />
+  );
+}
