+++ apps/hyperlocalise-web/src/app/[lang]/(authenticated)/org/[organizationSlug]/automations/_components/workspace-automation-editor.stories.tsx
@@ -209,7 +209,7 @@ export const CreateWithMemories: Story = {
     form: createMemoriesAutomationFormFixture(),
   },
   play: async ({ canvas }) => {
-    await expect(canvas.getByText("Use organization memory")).toBeInTheDocument();
+    await expect(canvas.getByText("Use workspace guideline")).toBeInTheDocument();
     await expect(canvas.getByRole("button", { name: "Manage" })).toBeInTheDocument();
     await expect(canvas.getByText("Allow memory updates")).toBeInTheDocument();
     await expect(canvas.getByText("3 tools")).toBeInTheDocument();
