+++ apps/hyperlocalise-web/src/app/[lang]/(authenticated)/org/[organizationSlug]/knowledge/_components/knowledge-page-view.stories.tsx
@@ -38,8 +38,8 @@ type Story = StoryObj<typeof meta>;
 
 export const Empty: Story = {
   play: async ({ canvas, args }) => {
-    await expect(canvas.getByRole("heading", { name: "Knowledge" })).toBeInTheDocument();
-    await expect(canvas.getByRole("heading", { name: "Upload knowledge" })).toBeInTheDocument();
+    await expect(canvas.getByRole("heading", { name: "Guideline" })).toBeInTheDocument();
+    await expect(canvas.getByRole("heading", { name: "Upload guideline" })).toBeInTheDocument();
     await expect(canvas.getByRole("button", { name: "Add Google Drive" })).toBeInTheDocument();
     await expect(canvas.getByRole("button", { name: "Markdown/Text" })).toBeInTheDocument();
     await expect(
@@ -61,7 +61,7 @@ export const WithMemory: Story = {
     }),
   },
   play: async ({ canvas }) => {
-    await expect(canvas.getByRole("heading", { name: "Knowledge" })).toBeInTheDocument();
+    await expect(canvas.getByRole("heading", { name: "Guideline" })).toBeInTheDocument();
     await expect(canvas.getByRole("heading", { name: "Global guidance" })).toBeInTheDocument();
     await expect(canvas.getByRole("button", { name: "Add sources" })).toBeInTheDocument();
     await expect(canvas.getByRole("button", { name: "History" })).toBeInTheDocument();
@@ -124,10 +124,10 @@ export const Loading: Story = {
     mode: "loading",
   },
   play: async ({ canvas }) => {
-    await expect(canvas.getByRole("heading", { name: "Knowledge" })).toBeInTheDocument();
+    await expect(canvas.getByRole("heading", { name: "Guideline" })).toBeInTheDocument();
     await expect(canvas.getByRole("status")).toBeInTheDocument();
     await expect(
-      canvas.queryByRole("heading", { name: "Upload knowledge" }),
+      canvas.queryByRole("heading", { name: "Upload guideline" }),
     ).not.toBeInTheDocument();
     await expect(
       canvas.queryByRole("heading", { name: "Global guidance" }),
