+++ apps/hyperlocalise-web/src/app/[lang]/(authenticated)/org/[organizationSlug]/knowledge/_components/knowledge-upload-section.messages.ts
@@ -16,9 +16,9 @@ import { defineMessages } from "react-intl";
 
 export const knowledgeUploadSectionMessages = defineMessages({
   title: {
-    defaultMessage: "Upload knowledge",
+    defaultMessage: "Upload guideline",
     id: "GWQDU8Wxgx",
-    description: "Heading for the knowledge upload empty state",
+    description: "Heading for the guideline upload empty state",
   },
   dropHint: {
     defaultMessage: "Drag & drop or {chooseFiles} to upload.",
