+++ apps/hyperlocalise-web/src/components/app-shell/app-shell-title.test.ts
@@ -38,7 +38,8 @@ describe("getAppShellTitle", () => {
     ["/org/acme/projects/proj_1/settings", "Settings"],
     ["/org/acme/my-work", "My Jobs"],
     ["/org/acme/my-jobs", "My Jobs"],
-    ["/org/acme/knowledge", "Knowledge"],
+    ["/org/acme/knowledge", "Guideline"],
+    ["/org/acme/ai-engine", "AI Engine"],
     ["/org/acme/domains", "Domains"],
     ["/org/acme/domains/ld_1", "ld_1"],
     ["/org/acme/glossaries", "Glossaries"],
