+++ apps/hyperlocalise-web/src/lib/flags/workos-adapter.test.ts
@@ -151,7 +151,7 @@ describe("workosAdapter", () => {
 const intl = getIntlShape("en") as IntlShape;
 
 describe("filterNavigationByWorkspaceFlags", () => {
-  it("removes Automations, Knowledge, and Domains when workspace flags are disabled", () => {
+  it("removes Automations, Guideline, and Domains when workspace flags are disabled", () => {
     const groups = buildGlobalNavigationGroups("acme", intl);
     const filtered = filterNavigationByWorkspaceFlags(groups, {
       automations: false,
@@ -164,13 +164,15 @@ describe("filterNavigationByWorkspaceFlags", () => {
     const itemLabels = filtered.flatMap((group) => group.items.map((item) => item.label));
 
     expect(itemLabels).not.toContain("Automations");
-    expect(itemLabels).not.toContain("Knowledge");
+    expect(itemLabels).not.toContain("Guideline");
     expect(itemLabels).toContain("Issues");
+    expect(itemLabels).toContain("New Request");
+    expect(itemLabels).toContain("AI Engine");
     expect(itemLabels).not.toContain("Domains");
     expect(itemLabels).toContain("Projects");
   });
 
-  it("keeps Automations, Knowledge, Issues, and Domains when workspace flags are enabled", () => {
+  it("keeps Automations, Guideline, Issues, and Domains when workspace flags are enabled", () => {
     const groups = buildGlobalNavigationGroups("acme", intl);
     const filtered = filterNavigationByWorkspaceFlags(groups, {
       automations: true,
@@ -183,8 +185,9 @@ describe("filterNavigationByWorkspaceFlags", () => {
     const itemLabels = filtered.flatMap((group) => group.items.map((item) => item.label));
 
     expect(itemLabels).toContain("Automations");
-    expect(itemLabels).toContain("Knowledge");
+    expect(itemLabels).toContain("Guideline");
     expect(itemLabels).toContain("Issues");
     expect(itemLabels).toContain("Domains");
+    expect(itemLabels).toContain("AI Engine");
   });
 });
